backend/requirements.txt
# ===========================================
# ScamShield Backend - Production Dependencies
# ===========================================

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# HTTP Client
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# AI Providers (FREE tiers)
groq>=0.4.2
google-generativeai>=0.3.0

# MongoDB (async)
motor>=3.3.2
beanie>=1.24.0
pymongo>=4.6.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# Google OAuth
authlib>=1.3.0
itsdangerous>=2.1.2

# Email
aiosmtplib>=3.0.0
email-validator>=2.1.0

# Rate Limiting
slowapi>=0.1.9

# System Monitoring
psutil>=5.9.0

# Logging
structlog>=24.1.0

# CORS
starlette>=0.35.0

# Testing (dev only)
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.25.0

